+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Unreleased
+### Added
+- `ArrowheadClient`
+  - Function `obtain_id` to obtain client ID using dummy service workaround.
+
 ## 0.1.0 - 2022-04-05
 ### Added
 - `ArrowheadInterface`